Finished functionality for direct posting

+++ bb.sh
@@ -597,15 +597,14 @@ $template_tags_line_header keep-this-tag-format, tags-are-optional, beware-with-
 EOF
     fi
 
-    exit 0
-
     chmod 600 "$TMPFILE"
 
     post_status="E"
     filename=""
+
     while [[ $post_status != "p" && $post_status != "P" ]]; do
         [[ -n $filename ]] && rm "$filename" # Delete the generated html file, if any
-        $EDITOR "$TMPFILE"
+        [[ ! $direct ]] && $EDITOR "$TMPFILE" || post_status="P" ## Direct avoids all prompts
         if [[ $fmt == md ]]; then
             html_from_md=$(markdown "$TMPFILE")
             parse_file "$html_from_md"
@@ -615,26 +614,28 @@ EOF
         fi
 
         chmod 644 "$filename"
-        [[ -n $preview_url ]] || preview_url=$global_url
-        echo "To preview the entry, open $preview_url/$filename in your browser"
+        if [[ ! $direct ]]; then
+            [[ -n $preview_url ]] || preview_url=$global_url
+            echo "To preview the entry, open $preview_url/$filename in your browser"
 
-        echo -n "[P]ost this entry, [E]dit again, [D]raft for later? (p/E/d) "
-        read -r post_status
-        if [[ $post_status == d || $post_status == D ]]; then
-            mkdir -p "drafts/"
-            chmod 700 "drafts/"
+            echo -n "[P]ost this entry, [E]dit again, [D]raft for later? (p/E/d) "
+            read -r post_status
+            if [[ $post_status == d || $post_status == D ]]; then
+                mkdir -p "drafts/"
+                chmod 700 "drafts/"
 
-            title=$(head -n 1 $TMPFILE)
-            [[ -n $convert_filename ]] && title=$(echo "$title" | eval "$convert_filename")
-            [[ -n $title ]] || title=$RANDOM
+                title=$(head -n 1 $TMPFILE)
+                [[ -n $convert_filename ]] && title=$(echo "$title" | eval "$convert_filename")
+                [[ -n $title ]] || title=$RANDOM
 
-            draft=drafts/$title.$fmt
-            mv "$TMPFILE" "$draft"
-            chmod 600 "$draft"
-            rm "$filename"
-            delete_includes
-            echo "Saved your draft as '$draft'"
-            exit
+                draft=drafts/$title.$fmt
+                mv "$TMPFILE" "$draft"
+                chmod 600 "$draft"
+                rm "$filename"
+                delete_includes
+                echo "Saved your draft as '$draft'"
+                exit
+            fi
         fi
     done
 
@@ -942,6 +943,7 @@ make_rss() {
 
 # generate headers, footers, etc
 create_includes() {
+    echo "Calling"
     {
         echo "<h1 class=\"nomargin\"><a class=\"ablack\" href=\"$global_url/$index_file\">$global_title</a></h1>" 
         echo "<div id=\"description\">$global_description</div>"
